Fix header files
